requirements.txt
fastapi
google-api-python-client
requests
google-cloud-translate
uvicorn
